split the exhaustiveness-checking logic to its own module
`check_match` is now left with its grab bag of random checks.
